squash merge py

+++ dump.py
@@ -1,4 +1,5 @@
 import os, sys
+import logging
 import numpy as np
 import tables
 
@@ -9,7 +10,12 @@ def parse_args(argv):
         'fields': ['OPENDAP','index_node','data_node','size','replica','version','retracted','_timestamp','_version_','checksum','checksum_type','_eva_ensemble_aggregation','_eva_variable_aggregation','_eva_no_frequency'],
         'frequency': 'frequency',
         'limit': None,
+        'logfile': None,
+        'loglevel': logging.INFO,
+        'node': '/files',
         'overwrite': False,
+        'overwrite_empty': False,
+        'prune': False,
         'step': 100000,
     }
 
@@ -31,8 +37,23 @@ def parse_args(argv):
         elif argv[position] == '-l' or argv[position] == '--limit':
             args['limit'] = int(argv[position+1])
             position+=2
+        elif argv[position] == '--logfile':
+            args['logfile'] = argv[position+1]
+            position+=2
+        elif argv[position] == '--loglevel':
+            args['loglevel'] = int(argv[position+1])
+            position+=2
+        elif argv[position] == '--node':
+            args['node'] = argv[position+1]
+            position+=2
         elif argv[position] == '--overwrite':
             args['overwrite'] = True
+            position+=1
+        elif argv[position] == '--overwrite-empty':
+            args['overwrite_empty'] = True
+            position+=1
+        elif argv[position] == '--prune':
+            args['prune'] = True
             position+=1
         elif argv[position] == '--step':
             args['step'] = argv[position+1]
@@ -44,49 +65,62 @@ def parse_args(argv):
     return args
 
 args = parse_args(sys.argv)
+logging.basicConfig(filename=args['logfile'], format='%(asctime)s - %(message)s', level=args['loglevel'])
 
 os.makedirs(args['dest'], exist_ok=True)
 
 f = tables.open_file(args['from'], 'r')
-t = f.get_node('/files')
+t = f.get_node(args['node'])
 
-step = args['step']
 s = set()
-for i in range(len(t)//step + 1):
-    s.update( t.colindexes['_eva_ensemble_aggregation'].read_sorted(i*step, i*step+step) )
+logging.debug('Loading unique IDs using a step of %d', args['step'])
+for i in range(len(t)//args['step'] + 1):
+    s.update( t.colindexes['_eva_ensemble_aggregation'].read_sorted(i*args['step'], i*args['step']+args['step']) )
+logging.debug('Loaded %d unique IDs', len(s))
 
 total = 0
 for agg in sorted(filter(lambda x: x != b'', s)):
     dest = os.path.join(args['dest'], agg.decode('utf-8'))
-    if os.path.exists(dest) and (not args['overwrite']) and os.stat(dest).st_size != 0:
+    logging.debug('Preparing "%s"', dest)
+    if os.path.exists(dest) and os.stat(dest).st_size == 0 and not args['overwrite_empty'] and not args['overwrite']:
+        logging.debug('Ignoring dataset "%s" with size 0', dest)
+        continue
+
+    if os.path.exists(dest) and not args['overwrite']:
+        logging.debug('Ignoring existing dataset "%s"', dest)
         continue
 
     # find current aggregation files and corresponding fxs
-    dataset = t.read_where('''(_eva_ensemble_aggregation == {}) & (frequency != b"fx")'''.format(agg))
-    if len(dataset) == 0:
-        continue
+    dataset = t.where('''(_eva_ensemble_aggregation == {}) & (frequency != b"fx")'''.format(agg))
 
-    fx_agg = dataset[0]['_eva_no_frequency']
-    fxs = t.read_where('''(_eva_no_frequency == {}) & ({} == b"fx")'''.format(fx_agg, args['frequency']))
-
-    dataset = np.concatenate([dataset, fxs], axis=0)
     fh = open(dest, 'w')
+    fxid = b''
     for match in dataset:
+        fxid = match['_eva_no_frequency']
         fields = ','.join( [match[field].decode('utf-8') for field in args['fields']] )
         if fields != '' and len(match['OPENDAP']) > 0:
             print(fields, file=fh)
+    if fxid != b'':
+        fxs = t.where('''(_eva_no_frequency == {}) & ({} == b"fx")'''.format(fxid, args['frequency']))
+        for match in fxs:
+            fields = ','.join( [match[field].decode('utf-8') for field in args['fields']] )
+            if fields != '' and len(match['OPENDAP']) > 0:
+                print(fields, file=fh)
     fh.close()
 
-    # If no rows have been written, remove file
-    if os.stat(dest).st_size != 0:
-        total += 1
-        print(dest)
-    else:
+    # Check if generated dataset is empty
+    total += 1
+    if os.stat(dest).st_size == 0 and args['prune']:
+        logging.debug('Removing empty dataset "%s"', dest)
         os.remove(dest)
+    else:
+        logging.info('Dumped dataset "%s"', dest)
+        print(dest)
 
     # Stop if reached user's limit
     if args['limit'] is not None:
         if total == args['limit']:
             break
 
+logging.info('Dumped %d datasets', total)
 f.close()
